blog-api/requirements.txt
fastapi
uvicorn
mypy
pydantic
SQLAlchemy
psycopg2 #psycopg2 for linux
python-dotenv
alembic
